first pass at converting LTDR_NDVI to use Dataset class
