+++ ltdr_ndvi/ingest/v4_yearly_raster_ingest.json
