+++ ltdr_ndvi/ingest/v5_monthly_raster_ingest.json
