+++ ltdr_ndvi/ingest/v5_yearly_raster_ingest.json
